# [chore] [exporter/splunkhec] Simplify the batching process