Rework how the batches are created and sent to the backend to make it easier to read the code.

Also, this change fixes an issue introduced by removing the additional `index` field of the `bufferState` struct in https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/21765: if a hec batch combining several OTel resources gets rejected with a retriable error, only the latest read resource/scope records were sent for retry. The fix is confirmed by an additional `TestPushLogsRetryableFailureMultipleResources` test.

This also significantly improves performance by removing a bunch of memory allocations in cases where the incoming data ends up being sent in several HEC batches.

Before:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
Benchmark_pushLogData_10_10_1024
Benchmark_pushLogData_10_10_1024-10    	   12033	     99060 ns/op	   90027 B/op	    1299 allocs/op
Benchmark_pushLogData_10_10_8K
Benchmark_pushLogData_10_10_8K-10      	   12072	     96328 ns/op	   90065 B/op	    1299 allocs/op
Benchmark_pushLogData_10_10_2M
Benchmark_pushLogData_10_10_2M-10      	   12306	     97606 ns/op	   90114 B/op	    1299 allocs/op
Benchmark_pushLogData_10_200_2M
Benchmark_pushLogData_10_200_2M-10     	     572	   2104811 ns/op	 1828672 B/op	   26003 allocs/op
Benchmark_pushLogData_100_200_2M
Benchmark_pushLogData_100_200_2M-10    	      56	  19257562 ns/op	18350930 B/op	  259916 allocs/op
Benchmark_pushLogData_100_200_5M
Benchmark_pushLogData_100_200_5M-10    	      55	  19130223 ns/op	18467195 B/op	  259916 allocs/op
PASS
```

After:

```
Benchmark_pushLogData_10_10_1024
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
Benchmark_pushLogData_10_10_1024
Benchmark_pushLogData_10_10_1024-10    	   34998	     34225 ns/op	   30894 B/op	     450 allocs/op
Benchmark_pushLogData_10_10_8K
Benchmark_pushLogData_10_10_8K-10      	   12380	     94922 ns/op	   87420 B/op	    1264 allocs/op
Benchmark_pushLogData_10_10_2M
Benchmark_pushLogData_10_10_2M-10      	   12411	     96700 ns/op	   90155 B/op	    1299 allocs/op
Benchmark_pushLogData_10_200_2M
Benchmark_pushLogData_10_200_2M-10     	     560	   2088985 ns/op	 1827450 B/op	   26003 allocs/op
Benchmark_pushLogData_100_200_2M
Benchmark_pushLogData_100_200_2M-10    	      70	  15065196 ns/op	14020531 B/op	  198538 allocs/op
Benchmark_pushLogData_100_200_5M
Benchmark_pushLogData_100_200_5M-10    	      55	  18660002 ns/op	18467047 B/op	  259916 allocs/op
PASS
```